Move options to their own file in pkg/client
This refactors options off to their own file in order to make each file
comprehensibly long (pkg/client/interfaces.go was getting rather large
with all the different types of options).
